backend/requirements.txt
flask
openai-whisper
yt-dlp
ffmpeg-python
flask-cors
